Backend/PythonWorkers/seer_orchestrator/requirements.txt
# /Backend/PythonWorkers/seer_orchestrator/requirements.txt
flask
kafka-python
google-generativeai
google-cloud-bigquery
werkzeug==2.2.2
PyYAML==6.0.1